Shenas/requirements.txt
Django==3.2.6
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
PyJWT==2.1.0
